adding exception handler and changing request mapping to getmapping in controller
