+++ .idea/libraries/libs.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="libs">
+    <CLASSES>
+      <root url="file://$PROJECT_DIR$/TourGuide/libs" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="file://$PROJECT_DIR$/TourGuide/libs" />
+    </SOURCES>
+    <jarDirectory url="file://$PROJECT_DIR$/TourGuide/libs" recursive="false" />
+    <jarDirectory url="file://$PROJECT_DIR$/TourGuide/libs" recursive="false" type="SOURCES" />
+  </library>
+</component>
