--- TourGuide/src/main/java/com/openclassrooms/tourguide/TourguideApplication.java
@@ -1,13 +1,0 @@
-package com.openclassrooms.tourguide;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TourguideApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TourguideApplication.class, args);
-	}
-
-}
