+++ TourGuide/src/main/java/com/phildev/tourguide/TourguideApplication.java
@@ -1,0 +1,13 @@
+package com.phildev.tourguide;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TourguideApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TourguideApplication.class, args);
+	}
+
+}
