+++ TourGuide/src/main/java/com/phildev/tourguide/config/TourGuideModule.java
@@ -1,0 +1,28 @@
+package com.phildev.tourguide.config;
+
+import com.phildev.tourguide.service.RewardsService;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import gpsUtil.GpsUtil;
+import rewardCentral.RewardCentral;
+
+@Configuration
+public class TourGuideModule {
+	
+	@Bean
+	public GpsUtil getGpsUtil() {
+		return new GpsUtil();
+	}
+	
+	@Bean
+	public RewardsService getRewardsService() {
+		return new RewardsService(getGpsUtil(), getRewardCentral());
+	}
+	
+	@Bean
+	public RewardCentral getRewardCentral() {
+		return new RewardCentral();
+	}
+	
+}
