+++ TourGuide/src/test/java/com/phildev/tourguide/TestTourGuideService.java
@@ -1,0 +1,128 @@
+package com.phildev.tourguide;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.util.List;
+import java.util.UUID;
+
+import com.phildev.tourguide.helper.InternalTestHelper;
+import com.phildev.tourguide.service.RewardsService;
+import com.phildev.tourguide.service.TourGuideService;
+import com.phildev.tourguide.user.User;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+
+import gpsUtil.GpsUtil;
+import gpsUtil.location.Attraction;
+import gpsUtil.location.VisitedLocation;
+import rewardCentral.RewardCentral;
+import tripPricer.Provider;
+
+public class TestTourGuideService {
+
+	@Test
+	public void getUserLocation() {
+		GpsUtil gpsUtil = new GpsUtil();
+		RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
+		InternalTestHelper.setInternalUserNumber(0);
+		TourGuideService tourGuideService = new TourGuideService(gpsUtil, rewardsService);
+
+		User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
+		VisitedLocation visitedLocation = tourGuideService.trackUserLocation(user);
+		tourGuideService.tracker.stopTracking();
+        assertEquals(visitedLocation.userId, user.getUserId());
+	}
+
+	@Test
+	public void addUser() {
+		GpsUtil gpsUtil = new GpsUtil();
+		RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
+		InternalTestHelper.setInternalUserNumber(0);
+		TourGuideService tourGuideService = new TourGuideService(gpsUtil, rewardsService);
+
+		User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
+		User user2 = new User(UUID.randomUUID(), "jon2", "000", "jon2@tourGuide.com");
+
+		tourGuideService.addUser(user);
+		tourGuideService.addUser(user2);
+
+		User retrievedUser = tourGuideService.getUser(user.getUserName());
+		User retrievedUser2 = tourGuideService.getUser(user2.getUserName());
+
+		tourGuideService.tracker.stopTracking();
+
+		assertEquals(user, retrievedUser);
+		assertEquals(user2, retrievedUser2);
+	}
+
+	@Test
+	public void getAllUsers() {
+		GpsUtil gpsUtil = new GpsUtil();
+		RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
+		InternalTestHelper.setInternalUserNumber(0);
+		TourGuideService tourGuideService = new TourGuideService(gpsUtil, rewardsService);
+
+		User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
+		User user2 = new User(UUID.randomUUID(), "jon2", "000", "jon2@tourGuide.com");
+
+		tourGuideService.addUser(user);
+		tourGuideService.addUser(user2);
+
+		List<User> allUsers = tourGuideService.getAllUsers();
+
+		tourGuideService.tracker.stopTracking();
+
+		assertTrue(allUsers.contains(user));
+		assertTrue(allUsers.contains(user2));
+	}
+
+	@Test
+	public void trackUser() {
+		GpsUtil gpsUtil = new GpsUtil();
+		RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
+		InternalTestHelper.setInternalUserNumber(0);
+		TourGuideService tourGuideService = new TourGuideService(gpsUtil, rewardsService);
+
+		User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
+		VisitedLocation visitedLocation = tourGuideService.trackUserLocation(user);
+
+		tourGuideService.tracker.stopTracking();
+
+		assertEquals(user.getUserId(), visitedLocation.userId);
+	}
+
+	@Disabled // Not yet implemented
+	@Test
+	public void getNearbyAttractions() {
+		GpsUtil gpsUtil = new GpsUtil();
+		RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
+		InternalTestHelper.setInternalUserNumber(0);
+		TourGuideService tourGuideService = new TourGuideService(gpsUtil, rewardsService);
+
+		User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
+		VisitedLocation visitedLocation = tourGuideService.trackUserLocation(user);
+
+		List<Attraction> attractions = tourGuideService.getNearByAttractions(visitedLocation);
+
+		tourGuideService.tracker.stopTracking();
+
+		assertEquals(5, attractions.size());
+	}
+
+	public void getTripDeals() {
+		GpsUtil gpsUtil = new GpsUtil();
+		RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
+		InternalTestHelper.setInternalUserNumber(0);
+		TourGuideService tourGuideService = new TourGuideService(gpsUtil, rewardsService);
+
+		User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
+
+		List<Provider> providers = tourGuideService.getTripDeals(user);
+
+		tourGuideService.tracker.stopTracking();
+
+		assertEquals(10, providers.size());
+	}
+
+}
