+++ TourGuide/src/test/java/com/phildev/tourguide/TourguideApplicationTests.java
@@ -1,0 +1,13 @@
+package com.phildev.tourguide;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TourguideApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
